accelsim: fix integer overflow in  and invalid signature for  override

+++ accelsim/accel-sim-framework-dev/gpu-simulator/gpgpu-sim/src/gpgpu-sim/gpu-sim.cc
@@ -2136,7 +2136,8 @@ void gpgpu_sim::perf_memcpy_to_gpu(size_t dst_start_addr, size_t count) {
     //== 0);
 
     for (unsigned counter = 0; counter < count; counter += 32) {
-      const unsigned wr_addr = dst_start_addr + counter;
+      // ROMAN: FIX
+      const size_t wr_addr = dst_start_addr + counter;
       addrdec_t raw_addr;
       mem_access_sector_mask_t mask;
       mask.set(wr_addr % 128 / 32);
